move XMLNS constants to xml-model module
