+++ library/xml-model/src/main/java/org/cru/godtools/xml/Constants.java
@@ -1,0 +1,9 @@
+package org.cru.godtools.xml;
+
+public final class Constants {
+    // XML namespaces
+    public static final String XMLNS_MANIFEST = "https://mobile-content-api.cru.org/xmlns/manifest";
+    public static final String XMLNS_ANALYTICS = "https://mobile-content-api.cru.org/xmlns/analytics";
+    public static final String XMLNS_TRACT = "https://mobile-content-api.cru.org/xmlns/tract";
+    public static final String XMLNS_CONTENT = "https://mobile-content-api.cru.org/xmlns/content";
+}
