+++ ui/tract-renderer/build.gradle
@@ -22,6 +22,7 @@ dependencies {
     implementation project(':library:model-events')
     implementation project(':library:model-loader')
     implementation project(':library:sync-tasks')
+    implementation project(':library:xml-model')
 
     implementation project(':ui:base')
