+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/Constants.java
@@ -1,12 +1,6 @@
 package org.cru.godtools.tract;
 
 public final class Constants {
-    // XML namespaces
-    public static final String XMLNS_MANIFEST = "https://mobile-content-api.cru.org/xmlns/manifest";
-    public static final String XMLNS_ANALYTICS = "https://mobile-content-api.cru.org/xmlns/analytics";
-    public static final String XMLNS_TRACT = "https://mobile-content-api.cru.org/xmlns/tract";
-    public static final String XMLNS_CONTENT = "https://mobile-content-api.cru.org/xmlns/content";
-
     // deep link uri params
     public static final String PARAM_USE_DEVICE_LANGUAGE = "useDeviceLanguage";
     public static final String PARAM_PRIMARY_LANGUAGE = "primaryLanguage";
