+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/model/Manifest.java
@@ -27,9 +27,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 
-import static org.cru.godtools.tract.Constants.XMLNS_MANIFEST;
 import static org.cru.godtools.tract.model.Utils.parseColor;
 import static org.cru.godtools.tract.model.Utils.parseScaleType;
+import static org.cru.godtools.xml.Constants.XMLNS_MANIFEST;
 
 public final class Manifest extends Base implements Styles {
     private static final String XML_MANIFEST = "manifest";
