+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/model/Text.java
@@ -17,8 +17,8 @@ import org.xmlpull.v1.XmlPullParserException;
 
 import java.io.IOException;
 
-import static org.cru.godtools.tract.Constants.XMLNS_CONTENT;
 import static org.cru.godtools.tract.model.Utils.parseColor;
+import static org.cru.godtools.xml.Constants.XMLNS_CONTENT;
 
 public final class Text extends Content {
     static final String XML_TEXT = "text";
